+++ src/components/reminders/ReminderList.js
@@ -4,14 +4,14 @@ import PropTypes from 'prop-types';
 import Icon from 'components/common/Icon';
 import Spacer from 'components/common/Spacer';
 import TaskTitle from 'components/tasks/common/TaskTitle';
+import withBusyCheck from 'containers/WithBusyCheck';
 import { useTaskReminderApi } from 'hooks/UseTaskReminderApi';
 import { useSettingsApi } from 'hooks/UseSettingsApi';
 import { showReminder } from 'utils/ReminderUtils';
 import { formatDate } from 'utils/SettingUtils';
 
-function ReminderList(props) {
-    const settingsApi = useSettingsApi();
-    const taskReminderApi = useTaskReminderApi(props.date);
+function ReminderList({ apis }) {
+    const { settingsApi, taskReminderApi } = apis;
 
     const [selectedTaskId, setSelectedTaskId] = useState([]);
     const [snoozeMenuVisible, setSnoozeMenuVisible] = useState(false);
@@ -147,7 +147,11 @@ function ReminderList(props) {
 }
 
 ReminderList.propTypes = {
+    apis: PropTypes.object.isRequired,
     date: PropTypes.string.isRequired
 };
 
-export default ReminderList;
+export default withBusyCheck(ReminderList, ({ date }) => ({
+    settingsApi: useSettingsApi(),
+    taskReminderApi: useTaskReminderApi(date)
+}));