+++ src/components/tasks/batch/ModalBatchAddTasksManager.js
@@ -1,12 +1,14 @@
 import React, { useRef } from 'react';
 import { Button, Form, Modal } from 'antd';
+import PropTypes from 'prop-types';
 import BatchAddTasksManager from 'components/tasks/batch/BatchAddTasksManager';
 import Icon from 'components/common/Icon';
 import Spacer from 'components/common/Spacer';
+import withBusyCheck from 'containers/WithBusyCheck';
 import { useAppApi } from 'hooks/UseAppApi';
 
-function ModalBatchAddTasksManager() {
-    const appApi = useAppApi();
+function ModalBatchAddTasksManager({ apis }) {
+    const { appApi } = apis;
 
     const [form] = Form.useForm();
     const managerRef = useRef();
@@ -39,4 +41,10 @@ function ModalBatchAddTasksManager() {
     );
 }
 
-export default ModalBatchAddTasksManager;
+ModalBatchAddTasksManager.propTypes = {
+    apis: PropTypes.object.isRequired
+};
+
+export default withBusyCheck(ModalBatchAddTasksManager, () => ({
+    appApi: useAppApi()
+}));