+++ src/data/DataFieldComponents.js
@@ -165,6 +165,11 @@ export function getFieldComponents(type, options) {
                                         props.onCommit();
                                     }
                                 }}
+                                onChange={value => {
+                                    if (props.onCommit && value === null) {
+                                        props.onCommit();
+                                    }
+                                }}
                                 format={dateFormat}
                                 {...removeExtraProps(props)} />
                         );
@@ -210,6 +215,11 @@ export function getFieldComponents(type, options) {
                                 onBlur={props.onCommit}
                                 onOpenChange={status => {
                                     if (props.onCommit && !status) {
+                                        props.onCommit();
+                                    }
+                                }}
+                                onChange={value => {
+                                    if (props.onCommit && value === null) {
                                         props.onCommit();
                                     }
                                 }}