+++ src/actions/publication/googlecal/GoogleCalAccountInfoActions.js
@@ -2,6 +2,7 @@ import { google } from 'googleapis';
 import { setPublicationData } from 'actions/PublicationActions';
 import { getClient } from 'actions/publication/googlecal/GoogleCalAuthorizationActions';
 import { getSettings } from 'selectors/SettingSelectors';
+import logger from 'utils/LogUtils';
 
 export function getGoogleCalAccountInfo() {
     return async (dispatch, getState) => {
@@ -16,7 +17,7 @@ export function getGoogleCalAccountInfo() {
 
         const result = await profile.userinfo.get();
 
-        console.debug('Get account info', result);
+        logger.debug('Get account info', result.data);
 
         await dispatch(setPublicationData('googlecal', {
             accountInfo: result.data