+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Improve synchronization performance
 * Fix clear extended date picker
+* Fix batch edit tasks
 
 # 1.1.0